requirements.txt
langchain-groq
duckduckgo-search
langchain_community
tiktoken
langchainhub
langchain
langgraph
python-dotenv